Split UHDM and UHDM-Surelog frontends
