+++ uhdm-plugin/uhdmastfrontend.cc
@@ -31,9 +31,6 @@
 #include <unistd.h>
 #endif
 
-#include "ErrorReporting/Report.h"
-#include "surelog.h"
-
 namespace UHDM {
 	extern void visit_object (vpiHandle obj_h, int indent, const char *relation, std::set<const BaseClass*>* visited, std::ostream& out, bool shallowVisit = false);
 }
@@ -52,135 +49,6 @@ static int
 get_line_num(void)
 {
 	return 1;
-}
-
-unsigned int executeCompilation(
-		int argc, const char** argv, bool diff_comp_mode, bool fileunit,
-		SURELOG::ErrorContainer::Stats* overallStats = NULL) {
-	bool success = true;
-	bool noFatalErrors = true;
-	unsigned int codedReturn = 0;
-	SURELOG::SymbolTable* symbolTable = new SURELOG::SymbolTable();
-	SURELOG::ErrorContainer* errors = new SURELOG::ErrorContainer(symbolTable);
-	SURELOG::CommandLineParser* clp = new SURELOG::CommandLineParser(
-			errors, symbolTable, diff_comp_mode, fileunit);
-	success = clp->parseCommandLine(argc, argv);
-	bool parseOnly = clp->parseOnly();
-	errors->printMessages(clp->muteStdout());
-	if (success && (!clp->help())) {
-		SURELOG::scompiler* compiler = SURELOG::start_compiler(clp);
-		if (!compiler) codedReturn |= 1;
-		SURELOG::shutdown_compiler(compiler);
-	}
-	SURELOG::ErrorContainer::Stats stats;
-	if (!clp->help()) {
-		stats = errors->getErrorStats();
-		if (overallStats) (*overallStats) += stats;
-		if (stats.nbFatal) codedReturn |= 1;
-		if (stats.nbSyntax) codedReturn |= 2;
-		// Only return non-zero for fatal and syntax errors
-		// if (stats.nbError)
-		//	codedReturn |= 4;
-	}
-	bool noFErrors = true;
-	if (!clp->help()) noFErrors = errors->printStats(stats, clp->muteStdout());
-	if (noFErrors == false) {
-		noFatalErrors = false;
-	}
-
-	std::string ext_command = clp->getExeCommand();
-	if (!ext_command.empty()) {
-		std::string directory = symbolTable->getSymbol(clp->getFullCompileDir());
-		std::string fileList = directory + "/file.lst";
-		std::string command = ext_command + " " + fileList;
-		int result = system(command.c_str());
-		codedReturn |= result;
-		std::cout << "Command result: " << result << std::endl;
-	}
-	clp->logFooter();
-	if (diff_comp_mode && fileunit) {
-		SURELOG::Report* report = new SURELOG::Report();
-		std::pair<bool, bool> results =
-				report->makeDiffCompUnitReport(clp, symbolTable);
-		success = results.first;
-		noFatalErrors = results.second;
-		delete report;
-	}
-	clp->cleanCache();	// only if -nocache
-	delete clp;
-	delete symbolTable;
-	delete errors;
-	if ((!noFatalErrors) || (!success)) codedReturn |= 1;
-	if (parseOnly)
-		return 0;
-	else
-		return codedReturn;
-}
-
-enum COMP_MODE {
-	NORMAL,
-	DIFF,
-	BATCH,
-};
-
-int run_surelog(int argc, const char** argv) {
-	SURELOG::Waiver::initWaivers();
-
-	unsigned int codedReturn = 0;
-	COMP_MODE mode = NORMAL;
-	bool nostdout = false;
-	std::string batchFile;
-	std::string diff_unit_opt = "-diffcompunit";
-	std::string parseonly_opt = "-parseonly";
-	std::string batch_opt = "-batch";
-	std::string nostdout_opt = "-nostdout";
-	for (int i = 1; i < argc; i++) {
-		if (parseonly_opt == argv[i]) {
-		} else if (diff_unit_opt == argv[i]) {
-			mode = DIFF;
-		} else if (batch_opt == argv[i]) {
-			batchFile = argv[i + 1];
-			i++;
-			mode = BATCH;
-		} else if (nostdout_opt == argv[i]) {
-			nostdout = true;
-		}
-	}
-
-	switch (mode) {
-		case DIFF: {
-#if (defined(_MSC_VER) || defined(__MINGW32__) || defined(__CYGWIN__))
-			// REVISIT: Windows doesn't have the concept of forks!
-			// Implement it sequentially for now and optimize it if this
-			// proves to be a bottleneck (preferably, implemented as a
-			// cross platform solution).
-			executeCompilation(argc, argv, true, false);
-			codedReturn = executeCompilation(argc, argv, true, true);
-#else
-			pid_t pid = fork();
-			if (pid == 0) {
-				// child process
-				executeCompilation(argc, argv, true, false);
-			} else if (pid > 0) {
-				// parent process
-				codedReturn = executeCompilation(argc, argv, true, true);
-			} else {
-				// fork failed
-				printf("fork() failed!\n");
-				return 1;
-			}
-#endif
-			break;
-		}
-		case NORMAL:
-			codedReturn = executeCompilation(argc, argv, false, false);
-			break;
-		case BATCH:
-			printf("Currently batch mode is not supported!\n");
-			return 1;
-	}
-
-	return codedReturn;
 }
 
 struct UhdmAstFrontend : public Frontend {
